+++ src/main/java/techcourse/myblog/domain/repository/ArticleRepository.java
@@ -1,4 +1,4 @@
-package techcourse.myblog.repository;
+package techcourse.myblog.domain.repository;
 
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
