+++ src/main/java/techcourse/myblog/domain/repository/UserRepository.java
@@ -1,13 +1,12 @@
-package techcourse.myblog.repository;
+package techcourse.myblog.domain.repository;
 
-import org.springframework.data.repository.CrudRepository;
+import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import techcourse.myblog.domain.User;
 
 import java.util.Optional;
 
 @Repository
-public interface UserRepository extends CrudRepository<User, Long> {
-
+public interface UserRepository extends JpaRepository<User, Long> {
     Optional<User> findByEmail(String email);
 }
