+++ src/test/java/techcourse/myblog/web/ArticleControllerTests.java
@@ -13,7 +13,7 @@ import org.springframework.util.LinkedMultiValueMap;
 import org.springframework.util.MultiValueMap;
 import techcourse.myblog.domain.Article;
 import techcourse.myblog.dto.ArticleDto;
-import techcourse.myblog.repository.ArticleRepository;
+import techcourse.myblog.domain.repository.ArticleRepository;
 
 import java.util.ArrayList;
 import java.util.List;
